docs(Label): implement Themed within documentation
